% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quartet_fct_display.R
\name{anscombe_plot}
\alias{anscombe_plot}
\title{Plot Anscombe}
\usage{
anscombe_plot(dataset, dataset_num = 1)
}
\arguments{
\item{dataset}{can be piped into}

\item{dataset_num}{numeric. which dataset from the quartet to plot}
}
\description{
Generate anscombe plot
}
