% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoRilla.R
\name{CI2GCV}
\alias{CI2GCV}
\title{CI2GCV}
\usage{
CI2GCV(CI, df, m, alpha, percent = FALSE)
}
\arguments{
\item{CI}{Confidence interval}

\item{df}{degrees of freedom (or number of data points)}

\item{m}{mean of the log-transformed samples}

\item{alpha}{Type I error rate (0.05 for 95% CI)}

\item{percent}{Is the GCV to be expressed as a percent? Defaults to FALSE.}
}
\value{
GCV
}
\description{
This function converts Confidence Interval to Geometric Coeficient of Variation
}
\examples{
CI2GCV()
}
\keyword{CI,}
\keyword{GCV}
