% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GCI}
\alias{GCI}
\title{GCI}
\usage{
GCI(x, alpha = 0.05)
}
\arguments{
\item{x}{Data points, as a vector-like object}

\item{alpha}{Type I error rate (default 0.05 for 95\% CI)}
}
\value{
GCI (list of two elements $lower, $upper)
}
\description{
This function calculates Geometric Confidence Interval
}
\examples{
GCI((9:20)*10) # Returns $lower 119.3551 $upper 165.9306
}
\keyword{GCI}
