% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GCV}
\alias{GCV}
\title{GCV}
\usage{
GCV(x, percent = FALSE)
}
\arguments{
\item{percent}{Is the GCV to be expressed as a percent? Defaults to FALSE.}

\item{@param}{x Data points, as a vector-like object}
}
\value{
GCV
}
\description{
This function calculates Geometric Coefficient of Variation
}
\examples{
GCV((9:20)*10) ### Returns 0.2636907
}
\keyword{GCV}
