% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoRilla.R
\name{GCV2CI}
\alias{GCV2CI}
\title{GCV2CI}
\usage{
GCV2CI(GCV, df, m, alpha, percent = FALSE)
}
\arguments{
\item{GCV}{Geometric Coefficient of Variation}

\item{df}{degrees of freedom (or number of data points)}

\item{m}{mean of the log-transformed samples}

\item{alpha}{Type I error rate (0.05 for 95% CI)}

\item{percent}{Is the GCV expressed as a percent? Defaults to FALSE.}
}
\value{
GSEM
}
\description{
This function converts Geometric Coefficient of Variation to Geometric SEM
}
\examples{
GCV2CI()
}
\keyword{GCV,}
\keyword{GSEM}
