% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GCV2GCI}
\alias{GCV2GCI}
\title{GCV2GCI}
\usage{
GCV2GCI(GCV, df, gmn, alpha = 0.05, percent = FALSE)
}
\arguments{
\item{GCV}{Geometric Coefficient of Variation}

\item{df}{degrees of freedom (or number of data points)}

\item{gmn}{geometric mean of the samples}

\item{alpha}{Type I error rate (default 0.05 for 95\% CI)}

\item{percent}{Is the GCV expressed as a percent? Defaults to FALSE.}
}
\value{
GSEM
}
\description{
This function converts Geometric Coefficient of Variation to Geometric SEM
}
\examples{
GCV2GCI()
}
\keyword{GCI}
\keyword{GCV,}
