% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GCV2GSD}
\alias{GCV2GSD}
\title{GCV2GSD}
\usage{
GCV2GSD(GCV, percent = FALSE)
}
\arguments{
\item{GCV}{Geometric Coefficient of Variation}

\item{percent}{Is the GCV expressed as a percent? Defaults to FALSE.}
}
\value{
GSD
}
\description{
This function converts Geometric Coefficient of Variation to Geometric SD
}
\examples{
GCV2GSD()
}
\keyword{GCV,}
\keyword{GSD}
