% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GCV2GSEM}
\alias{GCV2GSEM}
\title{GCV2GSEM}
\usage{
GCV2GSEM(GCV, n, percent = FALSE)
}
\arguments{
\item{GCV}{Geometric Coefficient of Variation}

\item{n}{degrees of freedom (or number of data points)}

\item{percent}{Is the GCV expressed as a percent? Defaults to FALSE.}
}
\value{
GSEM
}
\description{
This function converts Geometric Coefficient of Variation to Geometric SEM
}
\examples{
GCV2GSEM()
}
\keyword{GCV,}
\keyword{GSEM}
