% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GMEAN}
\alias{GMEAN}
\title{GMEAN}
\usage{
GMEAN(x)
}
\arguments{
\item{x}{Data points, as a vector-like object}
}
\value{
GMEAN
}
\description{
This function calculates Geometric Mean from a set of data
}
\examples{
GMEAN((9:20)*10) ## Returns 140.729
}
\keyword{GMEAN}
