% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GSEM}
\alias{GSEM}
\title{GSEM}
\usage{
GSEM(x)
}
\arguments{
\item{x}{Data points, as a vector-like object}
}
\value{
GSEM
}
\description{
This function calculates Geometric SEM from a set of data
}
\examples{
GSEM()
}
\keyword{GSEM}
