% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{kGCV}
\alias{kGCV}
\title{kGCV}
\usage{
kGCV(x, percent = FALSE)
}
\arguments{
\item{percent}{Is the GCV to be expressed as a percent? Defaults to FALSE.}

\item{@param}{x Data points, as a vector-like object}
}
\value{
kGCV
}
\description{
This function calculates Geometric Coefficient of Variation
using the formula of Kirkwood, TBL (1979) Biometrics, Vol. 35, pp. 908-909
}
\examples{
kGCV((9:20)*10) ### Returns 0.295986
}
\keyword{kGCV}
