% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoRilla.R
\name{kGCV2GSD}
\alias{kGCV2GSD}
\title{kGCV2GSD}
\usage{
kGCV2GSD(GCV, percent = FALSE)
}
\arguments{
\item{GCV}{Geometric Coefficient of Variation}

\item{percent}{Is the GCV expressed as a percent? Defaults to FALSE.}
}
\value{
GSD
}
\description{
This function converts Geometric Coefficient of Variation to Geometric SD,
using the formula of Kirkwood (1982)
}
\examples{
kGCV2GSD()
}
\keyword{GCV,}
\keyword{GSD}
