% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GCI2GCV}
\alias{GCI2GCV}
\title{GCI2GCV}
\usage{
GCI2GCV(GCI, df, gmn, alpha = 0.05, percent = FALSE)
}
\arguments{
\item{GCI}{Geometric Confidence interval}

\item{df}{degrees of freedom (or number of data points)}

\item{alpha}{Type I error rate (0.05 for 95\% CI)}

\item{percent}{Is the GCV to be expressed as a percent? Defaults to FALSE.}

\item{gm}{geometric mean of the samples}
}
\value{
GCV
}
\description{
This function converts Confidence Interval to Geometric Coeficient of Variation
}
\examples{
GCI2GCV()
}
\keyword{GCI,}
\keyword{GCV}
