% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoRilla.R
\name{GMEAN}
\alias{GMEAN}
\title{GMEAN}
\usage{
GMEAN(x)
}
\arguments{
\item{x}{Data points, as a vector-like object}
}
\value{
GMEAN
}
\description{
This function calculates Geometric Standard Deviation from a set of data
}
\examples{
GMEAN()
}
\keyword{GMEAN}
