% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoRilla.R
\name{GSD}
\alias{GSD}
\title{GSD}
\usage{
GSD(x)
}
\arguments{
\item{x}{Geometric Standard Deviation}
}
\value{
GSD
}
\description{
This function calculates Geometric Standard Deviation from a set of data
}
\examples{
GSD()
}
\keyword{GSD}
