% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GSD2GSEM}
\alias{GSD2GSEM}
\title{GSD2GSEM}
\usage{
GSD2GSEM(GSD, n)
}
\arguments{
\item{GSD}{Geometric Standard Deviation}

\item{n}{degrees of freedom (or number of data points)}
}
\value{
GSEM
}
\description{
This function converts Geometric Standard Deviation to Geometric SEM
}
\examples{
GSD2GSEM()
}
\keyword{GSD,}
\keyword{GSEM}
